Fix the hyperfine scalar and tensor matrix elements.
It turns out that the I_z*S_z term contributes to both the tensor _and_ the
scalar term, not just the tensor term.  To actually evaluate this term, it
needs to be rewritten in spherical tensor operator form as

I_z*S_z = 2/3 * T^2_0(I,S) + 1/3 T^0_0(I,S)

* note: T^0_0(I,S) = \vec{I}\cdot\vec{S}

Thus, the effective coefficient for the scalar term is (b + c / 3), and the
effective coefficient for the tensor term is 2 c / 3.