Completely fix hyperfine stark splitting (at least of n=0 f=0 and f=1 states)
There was a small typo in the code based on J. Chem Phys 71, 389 (1982) (the
"explicit wigner" code) -- namely, a nuclear spin term was missed in the phase
factor (-1)^t, resulting in a complex value for the phase factor when it should
have been real.  This code now reproduces the results of PRA 98, 032513 (2018)
as far as I can tell (the plots generated look equivalent).

The code based on J. Chem Phys 105, 7412 (1996) (the "analytical" code) still
doesn't exactly match as the splitting between m_f=0 and m_f=+-1 it produces is
too small (~1.13 MHz at 50 kV/cm whereas the explicit wigner code generates a
splitting of ~2.07 MHz at 50 kV/cm, in at least approximate agreement with the
2018 paper).