MS-22, MS-18 Re-increase nmax to improve accuracy==> greatly improves <d_z>
MS-22, MS-18 #comment see this commit's log

It turns out that nmax=4 is not an acceptable approximation when the devonshire
potential is enabled. Setting nmax=20 completely changes the spectrum, giving
results much more in-line with PRA 98, 032513 (2018).

The plots attached to MS-22 and MS-33 show a clear difference between the
vacuum and the Devonshire-confined systems.  Technically, the ground state
zero-applied-field expectation value of d_z is still 0, and the expectation
values of <d_x> and <d_y> also tend to be around 0, however, this is not
entirely unexpected, since nearby states tend to mix due to numerical precision
and round-off errors.  Additionally, the X- and Y-oriented states should be
degenerate with each other no matter what the applied electric field is, since
it is oriented along the z-axis.  When the applied electric field is zero, the
Z-oriented states will also be degenerate with the X- and Y-oriented states.
Note that as usual, numerics might prevent the degeneracy from being exact,
although this will not prevent mixing.

** Note that because of finite precision and accuracy, the states used are
not true eigenstates of the Hamiltonian.  Although they will almost be
composed of a single "true" eigenstate, there will be small components of
other eigenstates.

** At some point, I'll look into diagonalizing d_x, d_y, and d_z in each of the
degenerate subspaces/checking whether they commute with the Hamitonian in said
subspaces to further investigate this