Make code work on Linux with GCC 13.2
This isn't really a port since the code was designed to be portable from the
beginning (there were a few mistakes, and MSVC, as usual, is overly permissive
with code that shouldn't compile).  This is somewhat preliminary

Notes:
* right now, static linking is used for SpinlabHyperfineLib and libstdc++.  This
is because ubuntu 18.04's libstdc++ is too old to support C++20 for obvious
reasons.  I will investigate using RPATH to allow dynamic linking
* GCC didn't support std::format before GCC 13 (I think), so GCC 13 is required
to build this code.  I have started support for libfmt, but it doesnt fully
work yet -- needs more testing