Major additions to plot state coeffs, new baf_state python module
This includes new "pruned" and "superpruned" views of the states, a set of
matshow based plots, more debug outputs including the ability to print the
"qsq" expectation values, and total probability debug output

To do this, some of the old python-based calculation code has been added,
as the new baf_state module.  However, note that some of the Hamiltonian
calculation methods are blatantly wrong (for example, the hyperfine structure).
Thus attempting to use baf_state to calculate spectra is explicitly unsupported