Optimization: implement CUDA-based calculation of expectation values
This calculates expectation values (really matrix elements) for the entire
basis at once, which ends up being a speedup over doing each one individually
on the CPU, even though parts of the output are unused. Namely, the entire
matrix is calculated when only the diagonal is needed, but this is still a
substantial speedup over the old approach.  I may look into this in the future
if output_state_info becomes a bottleneck again.