Important bugfix and document HyperfineCalculator
The bug was the "4-byte version bug": the version was sometimes accidentally
written as four bytes for files of version aefdat_version::rawmat_okq2. This
is because the version was written without being being cast to a u16 first, so
sizeof(aefdat_version) bytes woud be written.  GCC and clang chose to use 4
bytes for aefdat_version, so two extra zero bytes were present where the flags
should be. Fortunately,the flags were never written as all-zero during the time
this bug was present (and also never will be in newer versions because of
FLAG_AVOID_ZERO_FLAGS_BUG). Thus, if "flags" is read as zero in a rawmat_okq2
file, then it was actually the accidental extra two bytes of version, so we
need to reread flags to get the correct value.